+++ mciutil/cli/__init__.py
@@ -2,6 +2,4 @@
 """
 Command line interface
 """
-__author__ = 'Anthony Delosa'
-__email__ = 'adelosa@gmail.com'
-__version__ = '0.4.1'
+