+++ mciutil/cli/common.py
@@ -8,8 +8,16 @@ import os
 import sys
 import logging
 import csv
+from pkg_resources import resource_filename
 
 LOGGER = logging.getLogger(__name__)
+# this module gets loaded by all the CLI programs so will emit message
+# for all cli programs.
+print("""!!! WARINING !!!
+mciutil module has been deprecated
+Please consider using module cardutil instead
+see https://cardutil.readthedocs.io
+""")
 
 
 def update_cli_progress(current_val, end_val, bar_length=20):
@@ -42,6 +50,13 @@ def add_source_format_arg(parser):
         help="encoding format of source file",
         choices=["ebcdic", "ascii"],
         default="ebcdic"
+    )
+
+    parser.add_argument(
+        "--no1014blocking",
+        help="do not use 1014 block format. Just vbs type record",
+        dest="no_1014_blocking",
+        action="store_true"
     )
 
 
@@ -79,12 +94,14 @@ def get_config_filename(config_filename):
     user_home_dir = os.path.expanduser("~")
 
     if os.path.isfile(current_dir + "/" + config_filename):
-        return current_dir + "/" + config_filename
+        config_filename = current_dir + "/" + config_filename
     elif os.path.isfile(user_home_dir + "/." + config_filename):
-        return user_home_dir + "/." + config_filename
+        config_filename = user_home_dir + "/." + config_filename
     else:
-        module_dir = os.path.dirname(os.path.abspath(__file__))
-        return module_dir + "/" + config_filename
+        module_dir = resource_filename("mciutil", "cli")
+        config_filename = module_dir + "/" + config_filename
+    LOGGER.info("Using {0} config file".format(config_filename))
+    return config_filename
 
 
 def add_to_csv(data_list, field_list, output_filename):
@@ -96,20 +113,40 @@ def add_to_csv(data_list, field_list, output_filename):
     :param output_filename: filename for output CSV file
     :return: None
     """
+    try:
+        instance_type = unicode
+        file_mode = "wb"
+    except NameError:
+        instance_type = str
+        file_mode = "w"
+
     filtered_data_list = filter_data_list(data_list, field_list)
 
-    with open(output_filename, "w") as output_file:
+    with open(output_filename, file_mode) as output_file:
         writer = csv.DictWriter(output_file,
                                 fieldnames=field_list,
                                 extrasaction="ignore",
                                 lineterminator="\n")
+
         # python 2.6 does not support writeheader() so skip
         if sys.version_info[0] == 2 and sys.version_info[1] == 6:
             pass
         else:
             writer.writeheader()
 
-        writer.writerows(filtered_data_list)
+        for item in filtered_data_list:
+            if file_mode == "w":
+                row = dict(
+                    (k, v.decode('latin1') if not isinstance(v, instance_type) else v)
+                    for k, v in item.items()
+                )
+            else:
+                row = dict(
+                    (k, v.encode('utf-8') if isinstance(v, instance_type) else v)
+                    for k, v in item.items()
+                )
+            writer.writerow(row)
+
     LOGGER.info("%s records written", len(data_list))
 
 
@@ -138,7 +175,7 @@ def filter_dictionary(dictionary, field_list):
     return_dictionary = {}
     for item in dictionary:
         if item in field_list:
-            return_dictionary[item] = dictionary[item].decode()
+            return_dictionary[item] = dictionary[item]
 
     return return_dictionary